Lazy load Database Accumulator
